@@ -1,24 +1,26 @@
 apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
-apply plugin: 'kotlin-android-extensions'
+apply plugin: 'kotlin-parcelize'
 apply plugin: 'kotlin-kapt'
 apply plugin: 'com.google.gms.google-services'
 apply plugin: 'com.google.firebase.crashlytics'
 apply plugin: 'com.google.firebase.firebase-perf'
 
 android {
-    compileSdkVersion 29
+    compileSdkVersion 32
+    buildFeatures {
+        viewBinding true
+    }
 
     defaultConfig {
         applicationId "id.fajarproject.roommovie"
         minSdkVersion 23
-        targetSdkVersion 29
+        targetSdkVersion 32
         versionCode 2
         versionName "1.0.1-build${getDate()}"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         multiDexEnabled true
         vectorDrawables.useSupportLibrary = true
-
     }
 
     buildTypes {
