@@ -2,6 +2,7 @@ package id.fajarproject.roommovie.ui.base
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
+import androidx.viewbinding.ViewBinding
 import id.fajarproject.roommovie.ui.widget.OnItemClickListener
 
 
@@ -9,15 +10,19 @@ import id.fajarproject.roommovie.ui.widget.OnItemClickListener
  * Create by Fajar Adi Prasetyo on 01/07/2020.
  */
 
-class AdapterHolder(itemView: View, onItemClickListener: OnItemClickListener?) : RecyclerView.ViewHolder(itemView),
+class AdapterHolder<B : ViewBinding>(var binding: B, onItemClickListener: OnItemClickListener?) :
+    RecyclerView.ViewHolder(binding.root),
     View.OnClickListener {
-    private var onItemClickListener : OnItemClickListener? = null
+    private var onItemClickListener: OnItemClickListener? = null
 
     override fun onClick(v: View?) {
-        onItemClickListener?.onItemClick(v,adapterPosition)
+        onItemClickListener?.onItemClick(v, adapterPosition)
     }
+
     init {
-        itemView.setOnClickListener(this)
+        binding.root.setOnClickListener(this)
         this.onItemClickListener = onItemClickListener
     }
+
 }
+
