@@ -2,10 +2,11 @@ package id.fajarproject.roommovie.ui.base
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
+import androidx.viewbinding.ViewBinding
 
 
 /**
  * Create by Fajar Adi Prasetyo on 03/07/2020.
  */
 
-class LoadingViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView)
+class LoadingViewHolder<B : ViewBinding>(viewBinding: B) : RecyclerView.ViewHolder(viewBinding.root)
