@@ -9,12 +9,12 @@ import androidx.core.view.ViewCompat
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import id.fajarproject.roommovie.R
+import id.fajarproject.roommovie.databinding.AdapterDetailVideoBinding
 import id.fajarproject.roommovie.models.PicturesItem
 import id.fajarproject.roommovie.ui.base.AdapterHolder
 import id.fajarproject.roommovie.ui.widget.OnItemClickListener
 import id.fajarproject.roommovie.util.Constant
 import id.fajarproject.roommovie.util.Util
-import kotlinx.android.synthetic.main.adapter_detail_video.view.*
 
 
 /**
@@ -24,7 +24,7 @@ class DetailBackdropsAdapter(
     var activity: Activity,
     private var list: MutableList<PicturesItem?>
 ) :
-    RecyclerView.Adapter<RecyclerView.ViewHolder>() {
+    RecyclerView.Adapter<AdapterHolder<AdapterDetailVideoBinding>>() {
 
     private var onItemClickListener : OnItemClickListener? = null
 
@@ -32,29 +32,27 @@ class DetailBackdropsAdapter(
         this.onItemClickListener = onItemClickListener
     }
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): AdapterHolder<AdapterDetailVideoBinding> {
         return AdapterHolder(
-            LayoutInflater.from(
-                parent.context
-            ).inflate(R.layout.adapter_detail_video, parent, false)
+            AdapterDetailVideoBinding.inflate(LayoutInflater.from(parent.context), parent, false)
             , this.onItemClickListener
         )
     }
 
     @SuppressLint("SetTextI18n")
     override fun onBindViewHolder(
-        holder: RecyclerView.ViewHolder,
+        holder: AdapterHolder<AdapterDetailVideoBinding>,
         position: Int
     ) {
-        Util.setViewPercents(activity, arrayOf(holder.itemView.viewMovie))
+        Util.setViewPercents(activity, arrayOf(holder.binding.viewMovie))
         val data = list[position] ?: PicturesItem()
 
         Glide.with(activity)
             .load(Constant.BASE_IMAGE + data.filePath)
             .error(R.drawable.ic_placeholder)
             .placeholder(Util.circleLoading(activity))
-            .into(holder.itemView.ivVideo)
-        holder.itemView.ivPlay.visibility   = View.GONE
+            .into(holder.binding.ivVideo)
+        holder.binding.ivPlay.visibility   = View.GONE
         val transitionName = activity.getString(R.string.transition_title,position)
         holder.itemView.tag = transitionName
         ViewCompat.setTransitionName(holder.itemView,transitionName)
