@@ -7,12 +7,12 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import id.fajarproject.roommovie.R
+import id.fajarproject.roommovie.databinding.AdapterDetailRecommendationBinding
 import id.fajarproject.roommovie.models.MovieItem
 import id.fajarproject.roommovie.ui.base.AdapterHolder
 import id.fajarproject.roommovie.ui.widget.OnItemClickListener
 import id.fajarproject.roommovie.util.Constant
 import id.fajarproject.roommovie.util.Util
-import kotlinx.android.synthetic.main.adapter_detail_recommendation.view.*
 
 /**
  * Create by Fajar Adi Prasetyo on 08/07/2020.
@@ -22,37 +22,41 @@ class DetailRecommendationsAdapter(
     var activity: Activity,
     private var list: MutableList<MovieItem?>
 ) :
-    RecyclerView.Adapter<RecyclerView.ViewHolder>() {
+    RecyclerView.Adapter<AdapterHolder<AdapterDetailRecommendationBinding>>() {
 
-    private var onItemClickListener : OnItemClickListener? = null
+    private var onItemClickListener: OnItemClickListener? = null
 
-    fun setOnItemClickListener(onItemClickListener: OnItemClickListener?){
+    fun setOnItemClickListener(onItemClickListener: OnItemClickListener?) {
         this.onItemClickListener = onItemClickListener
     }
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
+    override fun onCreateViewHolder(
+        parent: ViewGroup,
+        viewType: Int
+    ): AdapterHolder<AdapterDetailRecommendationBinding> {
         return AdapterHolder(
-            LayoutInflater.from(
-                parent.context
-            ).inflate(R.layout.adapter_detail_recommendation, parent, false)
-            , this.onItemClickListener
+            AdapterDetailRecommendationBinding.inflate(
+                LayoutInflater.from(parent.context),
+                parent,
+                false
+            ), this.onItemClickListener
         )
     }
 
     @SuppressLint("SetTextI18n")
     override fun onBindViewHolder(
-        holder: RecyclerView.ViewHolder,
+        holder: AdapterHolder<AdapterDetailRecommendationBinding>,
         position: Int
     ) {
-        Util.setViewPercents(activity, arrayOf(holder.itemView.viewMovie),0.8)
+        Util.setViewPercents(activity, arrayOf(holder.binding.viewMovie), 0.8)
         val data = list[position] ?: MovieItem()
-        holder.itemView.tvTitle.text    = data.title ?: data.name ?: ""
-        holder.itemView.tvRatting.text  = "${data.voteAverage?.times(10)} %"
+        holder.binding.tvTitle.text = data.title ?: data.name ?: ""
+        holder.binding.tvRatting.text = "${data.voteAverage?.times(10)} %"
         Glide.with(activity)
             .load(Constant.BASE_IMAGE + data.posterPath)
             .error(R.drawable.ic_placeholder)
             .placeholder(Util.circleLoading(activity))
-            .into(holder.itemView.ivMovie)
+            .into(holder.binding.ivMovie)
     }
 
     override fun getItemCount(): Int {
