@@ -7,12 +7,12 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import id.fajarproject.roommovie.R
+import id.fajarproject.roommovie.databinding.AdapterDetailVideoBinding
 import id.fajarproject.roommovie.models.VideosItem
 import id.fajarproject.roommovie.ui.base.AdapterHolder
 import id.fajarproject.roommovie.ui.widget.OnItemClickListener
 import id.fajarproject.roommovie.util.Constant
 import id.fajarproject.roommovie.util.Util
-import kotlinx.android.synthetic.main.adapter_detail_video.view.*
 
 /**
  * Create by Fajar Adi Prasetyo on 08/07/2020.
@@ -21,35 +21,39 @@ class DetailVideoAdapter(
     var activity: Activity,
     private var list: MutableList<VideosItem?>
 ) :
-    RecyclerView.Adapter<RecyclerView.ViewHolder>() {
+    RecyclerView.Adapter<AdapterHolder<AdapterDetailVideoBinding>>() {
 
-    private var onItemClickListener : OnItemClickListener? = null
+    private var onItemClickListener: OnItemClickListener? = null
 
-    fun setOnItemClickListener(onItemClickListener: OnItemClickListener?){
+    fun setOnItemClickListener(onItemClickListener: OnItemClickListener?) {
         this.onItemClickListener = onItemClickListener
     }
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
+    override fun onCreateViewHolder(
+        parent: ViewGroup,
+        viewType: Int
+    ): AdapterHolder<AdapterDetailVideoBinding> {
         return AdapterHolder(
-            LayoutInflater.from(
-                parent.context
-            ).inflate(R.layout.adapter_detail_video, parent, false)
-            , this.onItemClickListener
+            AdapterDetailVideoBinding.inflate(
+                LayoutInflater.from(parent.context),
+                parent,
+                false
+            ), this.onItemClickListener
         )
     }
 
     @SuppressLint("SetTextI18n")
     override fun onBindViewHolder(
-        holder: RecyclerView.ViewHolder,
+        holder: AdapterHolder<AdapterDetailVideoBinding>,
         position: Int
     ) {
-        Util.setViewPercents(activity, arrayOf(holder.itemView.viewMovie))
+        Util.setViewPercents(activity, arrayOf(holder.binding.viewMovie))
         val data = list[position] ?: VideosItem()
         Glide.with(activity)
             .load(Constant.BASE_THUMBNAIL + data.key + Constant.DEFAULT_QUALITY)
             .error(R.drawable.ic_placeholder)
             .placeholder(Util.circleLoading(activity))
-            .into(holder.itemView.ivVideo)
+            .into(holder.binding.ivVideo)
     }
 
     override fun getItemCount(): Int {
