@@ -10,6 +10,8 @@ import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.facebook.shimmer.Shimmer
 import id.fajarproject.roommovie.R
+import id.fajarproject.roommovie.databinding.FragmentMovieBinding
+import id.fajarproject.roommovie.databinding.ShimmerHomeBinding
 import id.fajarproject.roommovie.di.component.DaggerFragmentComponent
 import id.fajarproject.roommovie.di.module.FragmentModule
 import id.fajarproject.roommovie.models.MovieItem
@@ -19,8 +21,6 @@ import id.fajarproject.roommovie.ui.movieList.MovieListActivity
 import id.fajarproject.roommovie.ui.widget.OnItemClickListener
 import id.fajarproject.roommovie.util.Constant
 import id.fajarproject.roommovie.util.Util
-import kotlinx.android.synthetic.main.fragment_movie.*
-import kotlinx.android.synthetic.main.shimmer_home.*
 import javax.inject.Inject
 
 /**
@@ -28,9 +28,13 @@ import javax.inject.Inject
  */
 
 
-class MovieFragment : BaseFragment(),MovieContract.View {
+class MovieFragment : BaseFragment(), MovieContract.View {
 
-    @Inject lateinit var presenter : MovieContract.Presenter
+    @Inject
+    lateinit var presenter: MovieContract.Presenter
+
+    private lateinit var movieBinding: FragmentMovieBinding
+    private lateinit var shimmerHomeBinding: ShimmerHomeBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -40,9 +44,11 @@ class MovieFragment : BaseFragment(),MovieContract.View {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
+    ): View {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_movie, container, false)
+        movieBinding = FragmentMovieBinding.inflate(inflater, container, false)
+        shimmerHomeBinding = ShimmerHomeBinding.inflate(inflater, container, false)
+        return movieBinding.root
     }
 
     override fun onDestroy() {
@@ -52,90 +58,90 @@ class MovieFragment : BaseFragment(),MovieContract.View {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        presenter.attach<Fragment>(this,this)
+        presenter.attach<Fragment>(this, this)
         presenter.loadDataPopular()
         setUI()
     }
 
     override fun moveToActivity(status: String) {
-        val intent = Intent(requireContext(),MovieListActivity::class.java)
-        intent.putExtra(Constant.INTENT_STATUS,status)
+        val intent = Intent(requireContext(), MovieListActivity::class.java)
+        intent.putExtra(Constant.INTENT_STATUS, status)
         startActivity(intent)
     }
 
     override fun showDataPopularSuccess(list: MutableList<MovieItem?>) {
-        viewPopular.visibility      = View.VISIBLE
-        val layoutManager           = LinearLayoutManager(activity)
-        layoutManager.orientation   = LinearLayoutManager.HORIZONTAL
-        rvPopular.layoutManager     = layoutManager
-        val adapter                 = MovieAdapter(requireContext(),list)
-        rvPopular.adapter           = adapter
+        movieBinding.viewPopular.visibility = View.VISIBLE
+        val layoutManager = LinearLayoutManager(activity)
+        layoutManager.orientation = LinearLayoutManager.HORIZONTAL
+        movieBinding.rvPopular.layoutManager = layoutManager
+        val adapter = MovieAdapter(requireContext(), list)
+        movieBinding.rvPopular.adapter = adapter
         setItemClickListener(adapter)
     }
 
     override fun showDataPopularFailed(message: String) {
-        Log.e("ErrorPopular",message)
-        viewPopular.visibility      = View.GONE
+        Log.e("ErrorPopular", message)
+        movieBinding.viewPopular.visibility = View.GONE
     }
 
     override fun showDataNowPlayingSuccess(list: MutableList<MovieItem?>) {
-        viewNowPlaying.visibility   = View.VISIBLE
-        val layoutManager           = LinearLayoutManager(activity)
-        layoutManager.orientation   = LinearLayoutManager.HORIZONTAL
-        rvNowPlaying.layoutManager  = layoutManager
-        val adapter                 = MovieAdapter(requireContext(),list)
-        rvNowPlaying.adapter        = adapter
+        movieBinding.viewNowPlaying.visibility = View.VISIBLE
+        val layoutManager = LinearLayoutManager(activity)
+        layoutManager.orientation = LinearLayoutManager.HORIZONTAL
+        movieBinding.rvNowPlaying.layoutManager = layoutManager
+        val adapter = MovieAdapter(requireContext(), list)
+        movieBinding.rvNowPlaying.adapter = adapter
         setItemClickListener(adapter)
     }
 
     override fun showDataNowPlayingFailed(message: String) {
-        Log.e("ErrorNowPlaying",message)
-        viewNowPlaying.visibility = View.GONE
+        Log.e("ErrorNowPlaying", message)
+        movieBinding.viewNowPlaying.visibility = View.GONE
     }
 
     override fun showDataTopRatedSuccess(list: MutableList<MovieItem?>) {
-        viewTopRated.visibility   = View.VISIBLE
-        val layoutManager         = LinearLayoutManager(activity)
+        movieBinding.viewTopRated.visibility = View.VISIBLE
+        val layoutManager = LinearLayoutManager(activity)
         layoutManager.orientation = LinearLayoutManager.HORIZONTAL
-        rvTopRated.layoutManager  = layoutManager
-        val adapter               = MovieAdapter(requireContext(),list)
-        rvTopRated.adapter        = adapter
+        movieBinding.rvTopRated.layoutManager = layoutManager
+        val adapter = MovieAdapter(requireContext(), list)
+        movieBinding.rvTopRated.adapter = adapter
         setItemClickListener(adapter)
     }
 
     override fun showDataTopRatedFailed(message: String) {
-        Log.e("ErrorTopRated",message)
-        viewTopRated.visibility = View.GONE
+        Log.e("ErrorTopRated", message)
+        movieBinding.viewTopRated.visibility = View.GONE
     }
 
     override fun showDataUpcomingSuccess(list: MutableList<MovieItem?>) {
-        viewUpcoming.visibility   = View.VISIBLE
-        val layoutManager         = LinearLayoutManager(activity)
+        movieBinding.viewUpcoming.visibility = View.VISIBLE
+        val layoutManager = LinearLayoutManager(activity)
         layoutManager.orientation = LinearLayoutManager.HORIZONTAL
-        rvUpcoming.layoutManager  = layoutManager
-        val adapter               = MovieAdapter(requireContext(),list)
-        rvUpcoming.adapter        = adapter
+        movieBinding.rvUpcoming.layoutManager = layoutManager
+        val adapter = MovieAdapter(requireContext(), list)
+        movieBinding.rvUpcoming.adapter = adapter
         setItemClickListener(adapter)
     }
 
     override fun showDataUpcomingFailed(message: String) {
-        Log.e("ErrorUpcoming",message)
-        viewUpcoming.visibility = View.GONE
+        Log.e("ErrorUpcoming", message)
+        movieBinding.viewUpcoming.visibility = View.GONE
     }
 
     override fun showDataTrendingSuccess(list: MutableList<MovieItem?>) {
-        viewTrending.visibility   = View.VISIBLE
-        val layoutManager         = LinearLayoutManager(activity)
+        movieBinding.viewTrending.visibility = View.VISIBLE
+        val layoutManager = LinearLayoutManager(activity)
         layoutManager.orientation = LinearLayoutManager.HORIZONTAL
-        rvTrending.layoutManager  = layoutManager
-        val adapter               = MovieAdapter(requireContext(),list)
-        rvTrending.adapter        = adapter
+        movieBinding.rvTrending.layoutManager = layoutManager
+        val adapter = MovieAdapter(requireContext(), list)
+        movieBinding.rvTrending.adapter = adapter
         setItemClickListener(adapter)
     }
 
     override fun showDataTrendingFailed(message: String) {
-        Log.e("ErrorTrending",message)
-        viewTrending.visibility = View.GONE
+        Log.e("ErrorTrending", message)
+        movieBinding.viewTrending.visibility = View.GONE
     }
 
     override fun injectDependency() {
@@ -150,46 +156,65 @@ class MovieFragment : BaseFragment(),MovieContract.View {
     }
 
     override fun setUI() {
-        refreshLayout.isRefreshing = false
-        refreshLayout.setOnRefreshListener {
-            refreshLayout.isRefreshing = false
+        movieBinding.refreshLayout.isRefreshing = false
+        movieBinding.refreshLayout.setOnRefreshListener {
+            movieBinding.refreshLayout.isRefreshing = false
             presenter.loadDataPopular()
         }
-        cvPopular.setOnClickListener {
+        movieBinding.cvPopular.setOnClickListener {
             moveToActivity(getString(R.string.what_s_popular))
         }
-        cvNowPlaying.setOnClickListener {
+        movieBinding.cvNowPlaying.setOnClickListener {
             moveToActivity(getString(R.string.now_playing))
         }
-        cvTopRated.setOnClickListener {
+        movieBinding.cvTopRated.setOnClickListener {
             moveToActivity(getString(R.string.top_rated))
         }
-        cvUpcoming.setOnClickListener {
+        movieBinding.cvUpcoming.setOnClickListener {
             moveToActivity(getString(R.string.upcoming))
         }
-        cvTrending.setOnClickListener {
+        movieBinding.cvTrending.setOnClickListener {
             moveToActivity(getString(R.string.trending))
         }
-        Util.changeVisibility(arrayOf(llAiring,llOnTheAir),View.GONE)
-        Util.setViewPercents(requireContext(), arrayOf(viewShimmer1,viewShimmer2,viewShimmer3,viewShimmer4,viewShimmer5,viewShimmer6,viewShimmer7,viewShimmer8,viewShimmer9,viewShimmer10))
+        Util.changeVisibility(
+            arrayOf(shimmerHomeBinding.llAiring, shimmerHomeBinding.llOnTheAir),
+            View.GONE
+        )
+        Util.setViewPercents(
+            requireContext(),
+            arrayOf(
+                shimmerHomeBinding.viewShimmer1,
+                shimmerHomeBinding.viewShimmer2,
+                shimmerHomeBinding.viewShimmer3,
+                shimmerHomeBinding.viewShimmer4,
+                shimmerHomeBinding.viewShimmer5,
+                shimmerHomeBinding.viewShimmer6,
+                shimmerHomeBinding.viewShimmer7,
+                shimmerHomeBinding.viewShimmer8,
+                shimmerHomeBinding.viewShimmer9,
+                shimmerHomeBinding.viewShimmer10
+            )
+        )
     }
 
     override fun showLoading() {
-        scrollView.smoothScrollTo(0,0)
-        shimmerView.visibility  = View.VISIBLE
-        shimmerView.setShimmer(Shimmer.AlphaHighlightBuilder().setDuration(1150L).build())
-        shimmerView.startShimmer()
-        refreshLayout.visibility = View.GONE
+        shimmerHomeBinding.scrollView.smoothScrollTo(0, 0)
+        movieBinding.shimmerView.visibility = View.VISIBLE
+        movieBinding.shimmerView.setShimmer(
+            Shimmer.AlphaHighlightBuilder().setDuration(1150L).build()
+        )
+        movieBinding.shimmerView.startShimmer()
+        movieBinding.refreshLayout.visibility = View.GONE
     }
 
     override fun hideLoading() {
-        shimmerView.stopShimmer()
-        shimmerView.visibility      = View.GONE
-        refreshLayout.visibility    = View.VISIBLE
+        movieBinding.shimmerView.stopShimmer()
+        movieBinding.shimmerView.visibility = View.GONE
+        movieBinding.refreshLayout.visibility = View.VISIBLE
     }
 
-    override fun setItemClickListener(adapter: MovieAdapter){
-        adapter.setOnItemClickListener(object : OnItemClickListener{
+    override fun setItemClickListener(adapter: MovieAdapter) {
+        adapter.setOnItemClickListener(object : OnItemClickListener {
             override fun onItemClick(view: View?, position: Int) {
                 val item = adapter.getMovie(position)
                 item.id?.let {
@@ -199,9 +224,9 @@ class MovieFragment : BaseFragment(),MovieContract.View {
         })
     }
 
-    override fun moveToDetail(id : Int){
+    override fun moveToDetail(id: Int) {
         val intent = Intent(requireContext(), MovieDetailActivity::class.java)
-        intent.putExtra(Constant.idMovie,id)
+        intent.putExtra(Constant.idMovie, id)
         startActivity(intent)
     }
 }
