@@ -8,6 +8,8 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import id.fajarproject.roommovie.R
+import id.fajarproject.roommovie.databinding.AdapterLoadingBinding
+import id.fajarproject.roommovie.databinding.AdapterPeopleBinding
 import id.fajarproject.roommovie.models.people.KnownForItem
 import id.fajarproject.roommovie.models.people.PeopleItem
 import id.fajarproject.roommovie.ui.base.AdapterHolder
@@ -15,7 +17,6 @@ import id.fajarproject.roommovie.ui.base.LoadingViewHolder
 import id.fajarproject.roommovie.ui.widget.OnItemClickListener
 import id.fajarproject.roommovie.util.Constant
 import id.fajarproject.roommovie.util.Util
-import kotlinx.android.synthetic.main.adapter_people.view.*
 
 /**
  * Create by Fajar Adi Prasetyo on 03/07/2020.
@@ -28,24 +29,25 @@ class PeopleAdapter(
 
     private var isLoadingAdded = false
 
-    private var onItemClickListener : OnItemClickListener? = null
+    private var onItemClickListener: OnItemClickListener? = null
 
-    fun setOnItemClickListener(onItemClickListener: OnItemClickListener?){
+    fun setOnItemClickListener(onItemClickListener: OnItemClickListener?) {
         this.onItemClickListener = onItemClickListener
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
-        return if (viewType == Constant.VIEW_TYPE_ITEM){
+        return if (viewType == Constant.VIEW_TYPE_ITEM) {
             AdapterHolder(
-                LayoutInflater.from(
-                    parent.context
-                ).inflate(R.layout.adapter_people, parent, false)
-                , this.onItemClickListener
+                AdapterPeopleBinding.inflate(
+                    LayoutInflater.from(parent.context),
+                    parent,
+                    false
+                ), this.onItemClickListener
             )
-        }else {
+        } else {
             LoadingViewHolder(
-                LayoutInflater.from(parent.context).inflate(
-                    R.layout.adapter_loading,
+                AdapterLoadingBinding.inflate(
+                    LayoutInflater.from(parent.context),
                     parent,
                     false
                 )
@@ -59,14 +61,15 @@ class PeopleAdapter(
         position: Int
     ) {
         if (holder.itemViewType == Constant.VIEW_TYPE_ITEM) {
+            val binding = AdapterPeopleBinding.bind(holder.itemView)
             val data = list[position] ?: PeopleItem()
-            holder.itemView.tvName.text = data.name
-            holder.itemView.tvKnownFor.text = data.knownForDepartment + " • " + getKnowFor(data.knownFor)
+            binding.tvName.text = data.name
+            binding.tvKnownFor.text = data.knownForDepartment + " • " + getKnowFor(data.knownFor)
             Glide.with(activity)
                 .load(Constant.BASE_IMAGE + data.profilePath)
                 .error(R.drawable.ic_placeholder)
                 .placeholder(Util.circleLoading(activity))
-                .into(holder.itemView.ivProfile)
+                .into(binding.ivProfile)
         } else {
             Log.d("Loading", ".....")
         }
@@ -76,7 +79,7 @@ class PeopleAdapter(
         return list.size
     }
 
-    private fun getKnowFor(list : MutableList<KnownForItem?>?) : String{
+    private fun getKnowFor(list: MutableList<KnownForItem?>?): String {
         var knowFor = ""
         if (list != null) {
             for (i in list.indices) {
