@@ -7,11 +7,11 @@ import android.view.ViewGroup
 import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.RecyclerView
 import id.fajarproject.roommovie.R
+import id.fajarproject.roommovie.databinding.AdapterPeopleDetailCreditsBinding
 import id.fajarproject.roommovie.models.CreditsItem
 import id.fajarproject.roommovie.ui.base.AdapterHolder
 import id.fajarproject.roommovie.ui.widget.OnItemClickListener
 import id.fajarproject.roommovie.util.Util
-import kotlinx.android.synthetic.main.adapter_people_detail_credits.view.*
 
 /**
  * Create by Fajar Adi Prasetyo on 14/07/2020.
@@ -21,40 +21,47 @@ class PeopleDetailCreditsAdapter(
     var activity: Activity,
     private var list: MutableList<CreditsItem?>
 ) :
-    RecyclerView.Adapter<RecyclerView.ViewHolder>() {
+    RecyclerView.Adapter<AdapterHolder<AdapterPeopleDetailCreditsBinding>>() {
 
-    private var onItemClickListener : OnItemClickListener? = null
+    private var onItemClickListener: OnItemClickListener? = null
 
-    fun setOnItemClickListener(onItemClickListener: OnItemClickListener?){
+    fun setOnItemClickListener(onItemClickListener: OnItemClickListener?) {
         this.onItemClickListener = onItemClickListener
     }
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
+    override fun onCreateViewHolder(
+        parent: ViewGroup,
+        viewType: Int
+    ): AdapterHolder<AdapterPeopleDetailCreditsBinding> {
         return AdapterHolder(
-            LayoutInflater.from(
-                parent.context
-            ).inflate(R.layout.adapter_people_detail_credits, parent, false)
-            , this.onItemClickListener
+            AdapterPeopleDetailCreditsBinding.inflate(
+                LayoutInflater.from(parent.context),
+                parent,
+                false
+            ), this.onItemClickListener
         )
     }
 
     @SuppressLint("SetTextI18n")
     override fun onBindViewHolder(
-        holder: RecyclerView.ViewHolder,
+        holder: AdapterHolder<AdapterPeopleDetailCreditsBinding>,
         position: Int
     ) {
         val data = list[position] ?: CreditsItem()
-        holder.itemView.tvYear.text     =
+        holder.binding.tvYear.text =
             Util.convertDate(
                 data.releaseDate ?: data.firstAirDate ?: "",
                 "yyyy-MM-dd",
-                "yyyy")
+                "yyyy"
+            )
 
-        holder.itemView.tvTitle.text    = data.title ?: data.name
+        holder.binding.tvTitle.text = data.title ?: data.name
         data.episodeCount?.let {
-            Util.setSpannable(holder.itemView.tvTitle,
+            Util.setSpannable(
+                holder.binding.tvTitle,
                 " ($it ${activity.getString(R.string.episodes)})",
-                ContextCompat.getColor(activity,R.color.textColorSecondary))
+                ContextCompat.getColor(activity, R.color.textColorSecondary)
+            )
         }
         var info = ""
         data.character?.let {
@@ -62,7 +69,7 @@ class PeopleDetailCreditsAdapter(
         } ?: data.job?.let {
             info = "... $it"
         }
-        holder.itemView.tvInfo.text = info
+        holder.binding.tvInfo.text = info
     }
 
     override fun getItemCount(): Int {
