@@ -15,6 +15,7 @@ import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.facebook.shimmer.Shimmer
 import id.fajarproject.roommovie.R
+import id.fajarproject.roommovie.databinding.ActivityPictureBinding
 import id.fajarproject.roommovie.di.component.DaggerActivityComponent
 import id.fajarproject.roommovie.di.module.ActivityModule
 import id.fajarproject.roommovie.models.PicturesItem
@@ -24,40 +25,45 @@ import id.fajarproject.roommovie.ui.widget.GridSpacingItemDecoration
 import id.fajarproject.roommovie.ui.widget.OnItemClickListener
 import id.fajarproject.roommovie.util.Constant
 import id.fajarproject.roommovie.util.Util
-import kotlinx.android.synthetic.main.activity_picture.*
 import org.parceler.Parcels
 import javax.inject.Inject
 
-class PictureActivity : BaseActivity(),PictureContract.View {
+class PictureActivity : BaseActivity(), PictureContract.View {
 
-    @Inject lateinit var presenter : PictureContract.Presenter
-    lateinit var adapter : PictureAdapter
-    var list : MutableList<PicturesItem?> = arrayListOf()
+    @Inject
+    lateinit var presenter: PictureContract.Presenter
+    lateinit var adapter: PictureAdapter
+    var list: MutableList<PicturesItem?> = arrayListOf()
     private var reenterState: Bundle? = null
+    private lateinit var pictureBinding: ActivityPictureBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_picture)
+        pictureBinding = ActivityPictureBinding.inflate(layoutInflater)
+        setContentView(pictureBinding.root)
         injectDependency()
-        presenter.attach(this,this)
+        presenter.attach(this, this)
         setToolbar()
         setUI()
-        title       = intent.getStringExtra(Constant.title)
-        list        = Parcels.unwrap(intent.getParcelableExtra(Constant.dataPicture))
+        title = intent.getStringExtra(Constant.title)
+        list = Parcels.unwrap(intent.getParcelableExtra(Constant.dataPicture))
         showLoading()
-        if (list.isNotEmpty()){
+        if (list.isNotEmpty()) {
             showDataSuccess(list)
-        }else{
+        } else {
             checkData(false)
         }
-        ActivityCompat.setExitSharedElementCallback(this, presenter.exitElementCallback(reenterState,rvVideo))
+        ActivityCompat.setExitSharedElementCallback(
+            this,
+            presenter.exitElementCallback(reenterState, pictureBinding.rvVideo)
+        )
 
     }
 
     override fun showDataSuccess(list: MutableList<PicturesItem?>) {
         checkData(true)
-        rvVideo.layoutManager = GridLayoutManager(this,2)
-        rvVideo.addItemDecoration(
+        pictureBinding.rvVideo.layoutManager = GridLayoutManager(this, 2)
+        pictureBinding.rvVideo.addItemDecoration(
             GridSpacingItemDecoration(
                 activity,
                 2,
@@ -65,13 +71,13 @@ class PictureActivity : BaseActivity(),PictureContract.View {
                 true
             )
         )
-        rvVideo.itemAnimator    = DefaultItemAnimator()
-        adapter                 = PictureAdapter(this,list)
-        rvVideo.adapter         = adapter
+        pictureBinding.rvVideo.itemAnimator = DefaultItemAnimator()
+        adapter = PictureAdapter(this, list)
+        pictureBinding.rvVideo.adapter = adapter
         adapter.setOnItemClickListener(object : OnItemClickListener {
             override fun onItemClick(view: View?, position: Int) {
                 view?.let {
-                    showPreviewImage(it,position,list)
+                    showPreviewImage(it, position, list)
                 }
             }
         })
@@ -79,50 +85,53 @@ class PictureActivity : BaseActivity(),PictureContract.View {
     }
 
     override fun showDataFailed(message: String) {
-        Log.e("ErrorPicture",message)
+        Log.e("ErrorPicture", message)
         checkData(false)
     }
 
     override fun checkData(isShow: Boolean) {
         hideLoading()
-        if (isShow){
-            refreshLayout.visibility    = View.VISIBLE
-            noData.visibility           = View.GONE
-        }else{
-            refreshLayout.visibility    = View.GONE
-            noData.visibility           = View.VISIBLE
+        if (isShow) {
+            pictureBinding.refreshLayout.visibility = View.VISIBLE
+            pictureBinding.noData.visibility = View.GONE
+        } else {
+            pictureBinding.refreshLayout.visibility = View.GONE
+            pictureBinding.noData.visibility = View.VISIBLE
         }
     }
 
     override fun setViewBackToTop() {
-        rvVideo.addOnScrollListener(object : RecyclerView.OnScrollListener() {
+        pictureBinding.rvVideo.addOnScrollListener(object :
+            RecyclerView.OnScrollListener() {
             override fun onScrolled(recyclerView: RecyclerView, dx: Int, dy: Int) {
                 super.onScrolled(recyclerView, dx, dy)
-                val firstVisibleItem    = (recyclerView.layoutManager as GridLayoutManager).findFirstCompletelyVisibleItemPosition()
-                val lastVisibleItem     = (recyclerView.layoutManager as GridLayoutManager).findLastVisibleItemPosition()
-                if (lastVisibleItem < adapter.itemCount){
-                    btnBackToTop.show()
-                }else{
-                    btnBackToTop.hide()
+                val firstVisibleItem =
+                    (recyclerView.layoutManager as GridLayoutManager).findFirstCompletelyVisibleItemPosition()
+                val lastVisibleItem =
+                    (recyclerView.layoutManager as GridLayoutManager).findLastVisibleItemPosition()
+                if (lastVisibleItem < adapter.itemCount) {
+                    pictureBinding.btnBackToTop.show()
+                } else {
+                    pictureBinding.btnBackToTop.hide()
                 }
-                if (firstVisibleItem == 0){
-                    btnBackToTop.hide()
+                if (firstVisibleItem == 0) {
+                    pictureBinding.btnBackToTop.hide()
                 }
             }
         })
-        btnBackToTop.setOnClickListener {
-            rvVideo.smoothScrollToPosition(0)
+        pictureBinding.btnBackToTop.setOnClickListener {
+            pictureBinding.rvVideo.smoothScrollToPosition(0)
         }
     }
 
     override fun showPreviewImage(view: View, position: Int, data: MutableList<PicturesItem?>) {
         val intent = Intent(this, PreviewPictureActivity::class.java)
-        intent.putExtra(Constant.position,position)
+        intent.putExtra(Constant.position, position)
         intent.putExtra(Constant.dataPicture, Parcels.wrap(data))
-        val p1      = Pair.create(view, ViewCompat.getTransitionName(view))
-        val bundle  = ActivityOptionsCompat.makeSceneTransitionAnimation(this, p1).toBundle()
+        val p1 = Pair.create(view, ViewCompat.getTransitionName(view))
+        val bundle = ActivityOptionsCompat.makeSceneTransitionAnimation(this, p1).toBundle()
 
-        startActivity(intent,bundle)
+        startActivity(intent, bundle)
     }
 
     override fun clearReenterState() {
@@ -133,14 +142,17 @@ class PictureActivity : BaseActivity(),PictureContract.View {
         super.onActivityReenter(resultCode, data)
         reenterState = Bundle(data?.extras)
         reenterState?.let {
-            val startingPosition    = it.getInt(Constant.EXTRA_STARTING_ALBUM_POSITION)
-            val currentPosition     = it.getInt(Constant.EXTRA_CURRENT_ALBUM_POSITION)
-            if (startingPosition != currentPosition) rvVideo.scrollToPosition(currentPosition)
+            val startingPosition = it.getInt(Constant.EXTRA_STARTING_ALBUM_POSITION)
+            val currentPosition = it.getInt(Constant.EXTRA_CURRENT_ALBUM_POSITION)
+            if (startingPosition != currentPosition) pictureBinding.rvVideo.scrollToPosition(
+                currentPosition
+            )
             ActivityCompat.postponeEnterTransition(this)
 
-            rvVideo.viewTreeObserver.addOnPreDrawListener(object : ViewTreeObserver.OnPreDrawListener {
+            pictureBinding.rvVideo.viewTreeObserver.addOnPreDrawListener(object :
+                ViewTreeObserver.OnPreDrawListener {
                 override fun onPreDraw(): Boolean {
-                    rvVideo.viewTreeObserver.removeOnPreDrawListener(this)
+                    pictureBinding.rvVideo.viewTreeObserver.removeOnPreDrawListener(this)
                     ActivityCompat.startPostponedEnterTransition(this@PictureActivity)
                     return true
                 }
@@ -157,29 +169,38 @@ class PictureActivity : BaseActivity(),PictureContract.View {
     }
 
     override fun setToolbar() {
-        setSupportActionBar(toolbar)
+        setSupportActionBar(pictureBinding.toolbar)
         supportActionBar?.setDisplayHomeAsUpEnabled(true)
-        Util.setColorFilter(toolbar.navigationIcon!!, ContextCompat.getColor(this, R.color.iconColorPrimary))
+        Util.setColorFilter(
+            pictureBinding.toolbar.navigationIcon!!,
+            ContextCompat.getColor(this, R.color.iconColorPrimary)
+        )
     }
 
     override fun setUI() {
-        btnBackToTop.hide()
-        btnBackToTop.setOnClickListener { rvVideo.smoothScrollToPosition(0) }
-        refreshLayout.setOnRefreshListener {
-            refreshLayout.isRefreshing = false
+        pictureBinding.btnBackToTop.hide()
+        pictureBinding.btnBackToTop.setOnClickListener {
+            pictureBinding.rvVideo.smoothScrollToPosition(
+                0
+            )
+        }
+        pictureBinding.refreshLayout.setOnRefreshListener {
+            pictureBinding.refreshLayout.isRefreshing = false
         }
     }
 
     override fun showLoading() {
-        shimmerView.visibility  = View.VISIBLE
-        shimmerView.setShimmer(Shimmer.AlphaHighlightBuilder().setDuration(1150L).build())
-        shimmerView.startShimmer()
-        refreshLayout.visibility = View.GONE
+        pictureBinding.shimmerView.visibility = View.VISIBLE
+        pictureBinding.shimmerView.setShimmer(
+            Shimmer.AlphaHighlightBuilder().setDuration(1150L).build()
+        )
+        pictureBinding.shimmerView.startShimmer()
+        pictureBinding.refreshLayout.visibility = View.GONE
     }
 
     override fun hideLoading() {
-        shimmerView.stopShimmer()
-        shimmerView.visibility      = View.GONE
-        refreshLayout.visibility    = View.VISIBLE
+        pictureBinding.shimmerView.stopShimmer()
+        pictureBinding.shimmerView.visibility = View.GONE
+        pictureBinding.refreshLayout.visibility = View.VISIBLE
     }
 }
