@@ -9,13 +9,12 @@ import androidx.core.view.ViewCompat
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import id.fajarproject.roommovie.R
+import id.fajarproject.roommovie.databinding.AdapterImageBinding
 import id.fajarproject.roommovie.models.PicturesItem
 import id.fajarproject.roommovie.ui.base.AdapterHolder
 import id.fajarproject.roommovie.ui.widget.OnItemClickListener
 import id.fajarproject.roommovie.util.Constant
 import id.fajarproject.roommovie.util.Util
-import kotlinx.android.synthetic.main.adapter_detail_video.view.*
-
 
 /**
  * Create by Fajar Adi Prasetyo on 13/07/2020.
@@ -24,26 +23,30 @@ class PictureAdapter(
     var activity: Activity,
     private var list: MutableList<PicturesItem?>
 ) :
-    RecyclerView.Adapter<RecyclerView.ViewHolder>() {
+    RecyclerView.Adapter<AdapterHolder<AdapterImageBinding>>() {
 
-    private var onItemClickListener : OnItemClickListener? = null
+    private var onItemClickListener: OnItemClickListener? = null
 
-    fun setOnItemClickListener(onItemClickListener: OnItemClickListener?){
+    fun setOnItemClickListener(onItemClickListener: OnItemClickListener?) {
         this.onItemClickListener = onItemClickListener
     }
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
+    override fun onCreateViewHolder(
+        parent: ViewGroup,
+        viewType: Int
+    ): AdapterHolder<AdapterImageBinding> {
         return AdapterHolder(
-            LayoutInflater.from(
-                parent.context
-            ).inflate(R.layout.adapter_image, parent, false)
-            , this.onItemClickListener
+            AdapterImageBinding.inflate(
+                LayoutInflater.from(parent.context),
+                parent,
+                false
+            ), this.onItemClickListener
         )
     }
 
     @SuppressLint("SetTextI18n")
     override fun onBindViewHolder(
-        holder: RecyclerView.ViewHolder,
+        holder: AdapterHolder<AdapterImageBinding>,
         position: Int
     ) {
         val data = list[position] ?: PicturesItem()
@@ -51,11 +54,11 @@ class PictureAdapter(
             .load(Constant.BASE_IMAGE + data.filePath)
             .error(R.drawable.ic_placeholder)
             .placeholder(Util.circleLoading(activity))
-            .into(holder.itemView.ivVideo)
-        holder.itemView.ivPlay.visibility   = View.GONE
-        val transitionName  = activity.getString(R.string.transition_title,position)
+            .into(holder.binding.ivVideo)
+        holder.binding.ivPlay.visibility = View.GONE
+        val transitionName = activity.getString(R.string.transition_title, position)
         holder.itemView.tag = transitionName
-        ViewCompat.setTransitionName(holder.itemView,transitionName)
+        ViewCompat.setTransitionName(holder.itemView, transitionName)
     }
 
     override fun getItemCount(): Int {
