@@ -7,6 +7,8 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import id.fajarproject.roommovie.R
+import id.fajarproject.roommovie.databinding.AdapterLoadingBinding
+import id.fajarproject.roommovie.databinding.AdapterSearchBinding
 import id.fajarproject.roommovie.models.GenresItem
 import id.fajarproject.roommovie.models.MovieItem
 import id.fajarproject.roommovie.ui.base.AdapterHolder
@@ -14,7 +16,6 @@ import id.fajarproject.roommovie.ui.base.LoadingViewHolder
 import id.fajarproject.roommovie.ui.widget.OnItemClickListener
 import id.fajarproject.roommovie.util.Constant
 import id.fajarproject.roommovie.util.Util
-import kotlinx.android.synthetic.main.adapter_search.view.*
 
 
 /**
@@ -24,30 +25,31 @@ import kotlinx.android.synthetic.main.adapter_search.view.*
 class SearchAdapter(
     var activity: Activity,
     private var list: MutableList<MovieItem?>,
-    private val isMovie : Boolean
+    private val isMovie: Boolean
 ) :
     RecyclerView.Adapter<RecyclerView.ViewHolder>() {
 
     private var isLoadingAdded = false
 
-    private var onItemClickListener : OnItemClickListener? = null
+    private var onItemClickListener: OnItemClickListener? = null
 
-    fun setOnItemClickListener(onItemClickListener: OnItemClickListener?){
+    fun setOnItemClickListener(onItemClickListener: OnItemClickListener?) {
         this.onItemClickListener = onItemClickListener
     }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
-        return if (viewType == Constant.VIEW_TYPE_ITEM){
+        return if (viewType == Constant.VIEW_TYPE_ITEM) {
             AdapterHolder(
-                LayoutInflater.from(
-                    parent.context
-                ).inflate(R.layout.adapter_search, parent, false)
-                , this.onItemClickListener
+                AdapterSearchBinding.inflate(
+                    LayoutInflater.from(parent.context),
+                    parent,
+                    false
+                ), this.onItemClickListener
             )
-        }else {
+        } else {
             LoadingViewHolder(
-                LayoutInflater.from(parent.context).inflate(
-                    R.layout.adapter_loading,
+                AdapterLoadingBinding.inflate(
+                    LayoutInflater.from(parent.context),
                     parent,
                     false
                 )
@@ -60,27 +62,28 @@ class SearchAdapter(
         position: Int
     ) {
         if (holder.itemViewType == Constant.VIEW_TYPE_ITEM) {
+            val binding = AdapterSearchBinding.bind(holder.itemView)
             val data: MovieItem = list[position] ?: MovieItem()
-            holder.itemView.tvTitle.text = if (isMovie) data.title else data.name
+            binding.tvTitle.text = if (isMovie) data.title else data.name
             val dates = when {
                 data.releaseDate != null -> {
-                    Util.convertDate(data.releaseDate,"yyyy-MM-dd","dd MMMM yyyy")
+                    Util.convertDate(data.releaseDate, "yyyy-MM-dd", "dd MMMM yyyy")
                 }
                 data.firstAirDate != null -> {
-                    Util.convertDate(data.firstAirDate,"yyyy-MM-dd","dd MMMM yyyy")
+                    Util.convertDate(data.firstAirDate, "yyyy-MM-dd", "dd MMMM yyyy")
                 }
                 else -> {
                     "-"
                 }
             }
-            holder.itemView.tvDate.text     = dates
-            holder.itemView.tvRatting.text  = data.voteAverage.toString()
-            holder.itemView.tvGenre.text    = setGenre(data.genresIds ?: arrayListOf())
+            binding.tvDate.text = dates
+            binding.tvRatting.text = data.voteAverage.toString()
+            binding.tvGenre.text = setGenre(data.genresIds ?: arrayListOf())
             Glide.with(activity)
                 .load(Constant.BASE_IMAGE + data.posterPath)
                 .error(R.drawable.ic_placeholder)
                 .placeholder(Util.circleLoading(activity))
-                .into(holder.itemView.ivMovie)
+                .into(binding.ivMovie)
         } else {
             Log.d("Loading", ".....")
         }
@@ -120,12 +123,16 @@ class SearchAdapter(
         return if (position == list.size - 1 && isLoadingAdded) Constant.VIEW_TYPE_LOADING else Constant.VIEW_TYPE_ITEM
     }
 
-    private fun getGenre(id: Int) : String{
+    private fun getGenre(id: Int): String {
         var genre = ""
-        val list : MutableList<GenresItem?>? = if(isMovie) Util.getGenre(activity,Constant.genreMovie) else Util.getGenre(activity,Constant.genreTv)
+        val list: MutableList<GenresItem?>? =
+            if (isMovie) Util.getGenre(activity, Constant.genreMovie) else Util.getGenre(
+                activity,
+                Constant.genreTv
+            )
         if (list != null) {
-            for (data in list){
-                if (data?.id == id){
+            for (data in list) {
+                if (data?.id == id) {
                     genre = data.name ?: ""
                     break
                 }
@@ -134,14 +141,14 @@ class SearchAdapter(
         return genre
     }
 
-    private fun setGenre(list: MutableList<Int>) : String{
+    private fun setGenre(list: MutableList<Int>): String {
         var nameGenre = ""
-        for (i in list.indices){
+        for (i in list.indices) {
             val name = getGenre(list[i])
-            if (name.isNotEmpty()){
-                nameGenre += if (i == list.size - 1){
+            if (name.isNotEmpty()) {
+                nameGenre += if (i == list.size - 1) {
                     name
-                }else{
+                } else {
                     "$name \u2022 "
                 }
             }
