@@ -8,83 +8,87 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.facebook.shimmer.Shimmer
 import id.fajarproject.roommovie.R
+import id.fajarproject.roommovie.databinding.ActivitySeasonBinding
 import id.fajarproject.roommovie.di.component.DaggerActivityComponent
 import id.fajarproject.roommovie.di.module.ActivityModule
 import id.fajarproject.roommovie.models.SeasonsItem
 import id.fajarproject.roommovie.ui.base.BaseActivity
 import id.fajarproject.roommovie.util.Constant
 import id.fajarproject.roommovie.util.Util
-import kotlinx.android.synthetic.main.activity_season.*
 import org.parceler.Parcels
 
 /**
  * Create by Fajar Adi Prasetyo on 23/07/2020.
  */
 
-class SeasonActivity : BaseActivity(),SeasonContract.View {
+class SeasonActivity : BaseActivity(), SeasonContract.View {
 
     lateinit var adapter: SeasonAdapter
-    var list : MutableList<SeasonsItem?> = arrayListOf()
+    var list: MutableList<SeasonsItem?> = arrayListOf()
+    private lateinit var seasonBinding: ActivitySeasonBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_season)
+        seasonBinding = ActivitySeasonBinding.inflate(layoutInflater)
+        setContentView(seasonBinding.root)
         injectDependency()
         setToolbar()
         setUI()
-        title       = intent.getStringExtra(Constant.title)
-        list        = Parcels.unwrap(intent.getParcelableExtra(Constant.season))
+        title = intent.getStringExtra(Constant.title)
+        list = Parcels.unwrap(intent.getParcelableExtra(Constant.season))
         showLoading()
-        if (list.isNotEmpty()){
+        if (list.isNotEmpty()) {
             showDataSuccess(list)
-        }else{
+        } else {
             checkData(false)
         }
     }
 
     override fun showDataSuccess(list: MutableList<SeasonsItem?>) {
         checkData(true)
-        rvSeason.layoutManager  = LinearLayoutManager(this)
-        adapter                 = SeasonAdapter(this,list,title.toString())
-        rvSeason.adapter        = adapter
+        seasonBinding.rvSeason.layoutManager = LinearLayoutManager(this)
+        adapter = SeasonAdapter(this, list, title.toString())
+        seasonBinding.rvSeason.adapter = adapter
 
         setViewBackToTop()
     }
 
     override fun showDataFailed(message: String) {
-        Log.e("ErrorSeason",message)
+        Log.e("ErrorSeason", message)
         checkData(false)
     }
 
     override fun checkData(isShow: Boolean) {
         hideLoading()
-        if (isShow){
-            refreshLayout.visibility    = View.VISIBLE
-            noData.visibility           = View.GONE
-        }else{
-            refreshLayout.visibility    = View.GONE
-            noData.visibility           = View.VISIBLE
+        if (isShow) {
+            seasonBinding.refreshLayout.visibility = View.VISIBLE
+            seasonBinding.noData.visibility = View.GONE
+        } else {
+            seasonBinding.refreshLayout.visibility = View.GONE
+            seasonBinding.noData.visibility = View.VISIBLE
         }
     }
 
     override fun setViewBackToTop() {
-        rvSeason.addOnScrollListener(object : RecyclerView.OnScrollListener() {
+        seasonBinding.rvSeason.addOnScrollListener(object : RecyclerView.OnScrollListener() {
             override fun onScrolled(recyclerView: RecyclerView, dx: Int, dy: Int) {
                 super.onScrolled(recyclerView, dx, dy)
-                val firstVisibleItem    = (recyclerView.layoutManager as LinearLayoutManager).findFirstCompletelyVisibleItemPosition()
-                val lastVisibleItem     = (recyclerView.layoutManager as LinearLayoutManager).findLastVisibleItemPosition()
-                if (lastVisibleItem < adapter.itemCount){
-                    btnBackToTop.show()
-                }else{
-                    btnBackToTop.hide()
+                val firstVisibleItem =
+                    (recyclerView.layoutManager as LinearLayoutManager).findFirstCompletelyVisibleItemPosition()
+                val lastVisibleItem =
+                    (recyclerView.layoutManager as LinearLayoutManager).findLastVisibleItemPosition()
+                if (lastVisibleItem < adapter.itemCount) {
+                    seasonBinding.btnBackToTop.show()
+                } else {
+                    seasonBinding.btnBackToTop.hide()
                 }
-                if (firstVisibleItem == 0){
-                    btnBackToTop.hide()
+                if (firstVisibleItem == 0) {
+                    seasonBinding.btnBackToTop.hide()
                 }
             }
         })
-        btnBackToTop.setOnClickListener {
-            rvSeason.smoothScrollToPosition(0)
+        seasonBinding.btnBackToTop.setOnClickListener {
+            seasonBinding.rvSeason.smoothScrollToPosition(0)
         }
     }
 
@@ -97,29 +101,38 @@ class SeasonActivity : BaseActivity(),SeasonContract.View {
     }
 
     override fun setToolbar() {
-        setSupportActionBar(toolbar)
+        setSupportActionBar(seasonBinding.toolbar)
         supportActionBar?.setDisplayHomeAsUpEnabled(true)
-        Util.setColorFilter(toolbar.navigationIcon!!, ContextCompat.getColor(this, R.color.iconColorPrimary))
+        Util.setColorFilter(
+            seasonBinding.toolbar.navigationIcon!!,
+            ContextCompat.getColor(this, R.color.iconColorPrimary)
+        )
     }
 
     override fun setUI() {
-        btnBackToTop.hide()
-        btnBackToTop.setOnClickListener { rvSeason.smoothScrollToPosition(0) }
-        refreshLayout.setOnRefreshListener {
-            refreshLayout.isRefreshing = false
+        seasonBinding.btnBackToTop.hide()
+        seasonBinding.btnBackToTop.setOnClickListener {
+            seasonBinding.rvSeason.smoothScrollToPosition(
+                0
+            )
+        }
+        seasonBinding.refreshLayout.setOnRefreshListener {
+            seasonBinding.refreshLayout.isRefreshing = false
         }
     }
 
     override fun showLoading() {
-        shimmerView.visibility  = View.VISIBLE
-        shimmerView.setShimmer(Shimmer.AlphaHighlightBuilder().setDuration(1150L).build())
-        shimmerView.startShimmer()
-        refreshLayout.visibility = View.GONE
+        seasonBinding.shimmerView.visibility = View.VISIBLE
+        seasonBinding.shimmerView.setShimmer(
+            Shimmer.AlphaHighlightBuilder().setDuration(1150L).build()
+        )
+        seasonBinding.shimmerView.startShimmer()
+        seasonBinding.refreshLayout.visibility = View.GONE
     }
 
     override fun hideLoading() {
-        shimmerView.stopShimmer()
-        shimmerView.visibility      = View.GONE
-        refreshLayout.visibility    = View.VISIBLE
+        seasonBinding.shimmerView.stopShimmer()
+        seasonBinding.shimmerView.visibility = View.GONE
+        seasonBinding.refreshLayout.visibility = View.VISIBLE
     }
 }
