@@ -10,6 +10,8 @@ import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.facebook.shimmer.Shimmer
 import id.fajarproject.roommovie.R
+import id.fajarproject.roommovie.databinding.FragmentTvBinding
+import id.fajarproject.roommovie.databinding.ShimmerHomeBinding
 import id.fajarproject.roommovie.di.component.DaggerFragmentComponent
 import id.fajarproject.roommovie.di.module.FragmentModule
 import id.fajarproject.roommovie.models.MovieItem
@@ -20,15 +22,17 @@ import id.fajarproject.roommovie.ui.widget.OnItemClickListener
 import id.fajarproject.roommovie.util.Constant
 import id.fajarproject.roommovie.util.Util
 import javax.inject.Inject
-import kotlinx.android.synthetic.main.fragment_tv.*
-import kotlinx.android.synthetic.main.shimmer_home.*
 
 /**
  * Create by Fajar Adi Prasetyo on 01/07/2020.
  */
-class TvFragment : BaseFragment() ,TvContract.View {
+class TvFragment : BaseFragment(), TvContract.View {
 
-    @Inject lateinit var presenter: TvContract.Presenter
+    @Inject
+    lateinit var presenter: TvContract.Presenter
+
+    private lateinit var tvBinding: FragmentTvBinding
+    private lateinit var shimmerHomeBinding: ShimmerHomeBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -38,9 +42,11 @@ class TvFragment : BaseFragment() ,TvContract.View {
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
+    ): View {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_tv, container, false)
+        tvBinding = FragmentTvBinding.inflate(inflater, container, false)
+        shimmerHomeBinding = ShimmerHomeBinding.inflate(inflater, container, false)
+        return tvBinding.root
     }
 
     override fun onDestroy() {
@@ -50,75 +56,75 @@ class TvFragment : BaseFragment() ,TvContract.View {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        presenter.attach<Fragment>(this,this)
+        presenter.attach<Fragment>(this, this)
         presenter.loadDataAiringToday()
         setUI()
     }
 
     override fun moveToActivity(status: String) {
-        val intent = Intent(requireContext(),TvListActivity::class.java)
-        intent.putExtra(Constant.INTENT_STATUS,status)
+        val intent = Intent(requireContext(), TvListActivity::class.java)
+        intent.putExtra(Constant.INTENT_STATUS, status)
         startActivity(intent)
     }
 
     override fun showDataPopularSuccess(list: MutableList<MovieItem?>) {
-        viewPopular.visibility      = View.VISIBLE
-        val layoutManager           = LinearLayoutManager(activity)
-        layoutManager.orientation   = LinearLayoutManager.HORIZONTAL
-        rvPopular.layoutManager     = layoutManager
-        val adapter                 = TvAdapter(requireContext(),list)
-        rvPopular.adapter           = adapter
+        tvBinding.viewPopular.visibility = View.VISIBLE
+        val layoutManager = LinearLayoutManager(activity)
+        layoutManager.orientation = LinearLayoutManager.HORIZONTAL
+        tvBinding.rvPopular.layoutManager = layoutManager
+        val adapter = TvAdapter(requireContext(), list)
+        tvBinding.rvPopular.adapter = adapter
         setItemClickListener(adapter)
     }
 
     override fun showDataPopularFailed(message: String) {
-        Log.e("ErrorPopular",message)
-        viewPopular.visibility      = View.GONE
+        Log.e("ErrorPopular", message)
+        tvBinding.viewPopular.visibility = View.GONE
     }
 
     override fun showDataAiringTodaySuccess(list: MutableList<MovieItem?>) {
-        viewNowPlaying.visibility       = View.VISIBLE
-        val layoutManager               = LinearLayoutManager(activity)
-        layoutManager.orientation       = LinearLayoutManager.HORIZONTAL
-        rvNowPlaying.layoutManager      = layoutManager
-        val adapter                     = TvAdapter(requireContext(),list)
-        rvNowPlaying.adapter            = adapter
+        tvBinding.viewNowPlaying.visibility = View.VISIBLE
+        val layoutManager = LinearLayoutManager(activity)
+        layoutManager.orientation = LinearLayoutManager.HORIZONTAL
+        tvBinding.rvNowPlaying.layoutManager = layoutManager
+        val adapter = TvAdapter(requireContext(), list)
+        tvBinding.rvNowPlaying.adapter = adapter
         setItemClickListener(adapter)
     }
 
     override fun showDataAiringTodayFailed(message: String) {
-        Log.e("ErrorAiringToday",message)
-        viewNowPlaying.visibility      = View.GONE
+        Log.e("ErrorAiringToday", message)
+        tvBinding.viewNowPlaying.visibility = View.GONE
     }
 
     override fun showDataTopRatedSuccess(list: MutableList<MovieItem?>) {
-        viewTopRated.visibility      = View.VISIBLE
-        val layoutManager            = LinearLayoutManager(activity)
-        layoutManager.orientation    = LinearLayoutManager.HORIZONTAL
-        rvTopRated.layoutManager     = layoutManager
-        val adapter                  = TvAdapter(requireContext(),list)
-        rvTopRated.adapter           = adapter
+        tvBinding.viewTopRated.visibility = View.VISIBLE
+        val layoutManager = LinearLayoutManager(activity)
+        layoutManager.orientation = LinearLayoutManager.HORIZONTAL
+        tvBinding.rvTopRated.layoutManager = layoutManager
+        val adapter = TvAdapter(requireContext(), list)
+        tvBinding.rvTopRated.adapter = adapter
         setItemClickListener(adapter)
     }
 
     override fun showDataTopRatedFailed(message: String) {
-        Log.e("ErrorTopRated",message)
-        viewTopRated.visibility      = View.GONE
+        Log.e("ErrorTopRated", message)
+        tvBinding.viewTopRated.visibility = View.GONE
     }
 
     override fun showDataOnTheAirSuccess(list: MutableList<MovieItem?>) {
-        viewOnTheAir.visibility     = View.VISIBLE
-        val layoutManager           = LinearLayoutManager(activity)
-        layoutManager.orientation   = LinearLayoutManager.HORIZONTAL
-        rvOnTheAir.layoutManager    = layoutManager
-        val adapter                 = TvAdapter(requireContext(),list)
-        rvOnTheAir.adapter          = adapter
+        tvBinding.viewOnTheAir.visibility = View.VISIBLE
+        val layoutManager = LinearLayoutManager(activity)
+        layoutManager.orientation = LinearLayoutManager.HORIZONTAL
+        tvBinding.rvOnTheAir.layoutManager = layoutManager
+        val adapter = TvAdapter(requireContext(), list)
+        tvBinding.rvOnTheAir.adapter = adapter
         setItemClickListener(adapter)
     }
 
     override fun showDataOnTheAirFailed(message: String) {
-        Log.e("ErrorOnTheAir",message)
-        viewOnTheAir.visibility      = View.GONE
+        Log.e("ErrorOnTheAir", message)
+        tvBinding.viewOnTheAir.visibility = View.GONE
     }
 
     override fun setItemClickListener(adapter: TvAdapter) {
@@ -132,9 +138,9 @@ class TvFragment : BaseFragment() ,TvContract.View {
         })
     }
 
-    override fun moveToDetail(id : Int){
+    override fun moveToDetail(id: Int) {
         val intent = Intent(requireContext(), TvDetailActivity::class.java)
-        intent.putExtra(Constant.idMovie,id)
+        intent.putExtra(Constant.idMovie, id)
         startActivity(intent)
     }
 
@@ -150,40 +156,59 @@ class TvFragment : BaseFragment() ,TvContract.View {
     }
 
     override fun setUI() {
-        refreshLayout.isRefreshing = false
-        refreshLayout.setOnRefreshListener {
-            refreshLayout.isRefreshing = false
+        tvBinding.refreshLayout.isRefreshing = false
+        tvBinding.refreshLayout.setOnRefreshListener {
+            tvBinding.refreshLayout.isRefreshing = false
             presenter.loadDataAiringToday()
         }
-        cvOnTheAir.setOnClickListener {
+        tvBinding.cvOnTheAir.setOnClickListener {
             moveToActivity(getString(R.string.tv_on_the_air))
         }
-        cvNowPlaying.setOnClickListener {
+        tvBinding.cvNowPlaying.setOnClickListener {
             moveToActivity(getString(R.string.tv_airing_today))
         }
-        cvTopRated.setOnClickListener {
+        tvBinding.cvTopRated.setOnClickListener {
             moveToActivity(getString(R.string.top_rated))
         }
-        cvPopular.setOnClickListener {
+        tvBinding.cvPopular.setOnClickListener {
             moveToActivity(getString(R.string.what_s_popular))
         }
 
-        Util.changeVisibility(arrayOf(llNowPlaying,llTopRated),View.GONE)
-        Util.setViewPercents(requireContext(), arrayOf(viewShimmer1,viewShimmer2,viewShimmer3,viewShimmer4,viewShimmer5,viewShimmer6,viewShimmer7,viewShimmer8,viewShimmer9,viewShimmer10))
+        Util.changeVisibility(
+            arrayOf(
+                shimmerHomeBinding.llNowPlaying,
+                shimmerHomeBinding.llTopRated
+            ), View.GONE
+        )
+        Util.setViewPercents(
+            requireContext(),
+            arrayOf(
+                shimmerHomeBinding.viewShimmer1,
+                shimmerHomeBinding.viewShimmer2,
+                shimmerHomeBinding.viewShimmer3,
+                shimmerHomeBinding.viewShimmer4,
+                shimmerHomeBinding.viewShimmer5,
+                shimmerHomeBinding.viewShimmer6,
+                shimmerHomeBinding.viewShimmer7,
+                shimmerHomeBinding.viewShimmer8,
+                shimmerHomeBinding.viewShimmer9,
+                shimmerHomeBinding.viewShimmer10
+            )
+        )
     }
 
     override fun showLoading() {
-        scrollView.smoothScrollTo(0,0)
-        shimmerView.visibility  = View.VISIBLE
-        shimmerView.setShimmer(Shimmer.AlphaHighlightBuilder().setDuration(1150L).build())
-        shimmerView.startShimmer()
-        refreshLayout.visibility = View.GONE
+        shimmerHomeBinding.scrollView.smoothScrollTo(0, 0)
+        tvBinding.shimmerView.visibility = View.VISIBLE
+        tvBinding.shimmerView.setShimmer(Shimmer.AlphaHighlightBuilder().setDuration(1150L).build())
+        tvBinding.shimmerView.startShimmer()
+        tvBinding.refreshLayout.visibility = View.GONE
     }
 
     override fun hideLoading() {
-        shimmerView.stopShimmer()
-        shimmerView.visibility      = View.GONE
-        refreshLayout.visibility    = View.VISIBLE
+        tvBinding.shimmerView.stopShimmer()
+        tvBinding.shimmerView.visibility = View.GONE
+        tvBinding.refreshLayout.visibility = View.VISIBLE
     }
 
 }
