@@ -11,6 +11,7 @@ import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.GridLayoutManager
 import com.facebook.shimmer.Shimmer
 import id.fajarproject.roommovie.R
+import id.fajarproject.roommovie.databinding.ActivityTvListBinding
 import id.fajarproject.roommovie.di.component.DaggerActivityComponent
 import id.fajarproject.roommovie.di.module.ActivityModule
 import id.fajarproject.roommovie.models.MovieItem
@@ -20,15 +21,16 @@ import id.fajarproject.roommovie.ui.widget.OnItemClickListener
 import id.fajarproject.roommovie.util.Constant
 import id.fajarproject.roommovie.util.PaginationScrollListener
 import id.fajarproject.roommovie.util.Util
-import kotlinx.android.synthetic.main.activity_tv_list.*
 import javax.inject.Inject
 
-class TvListActivity : BaseActivity() ,TvListContract.View{
+class TvListActivity : BaseActivity(), TvListContract.View {
 
     @Inject
     lateinit var presenter: TvListContract.Presenter
     lateinit var layoutManager: GridLayoutManager
     lateinit var adapter: TvListAdapter
+    private lateinit var tvListBinding: ActivityTvListBinding
+
     var isLoading = false
     var isLastPage = false
 
@@ -36,13 +38,14 @@ class TvListActivity : BaseActivity() ,TvListContract.View{
     var currentPage = 1
     var limit = 20
 
-    private var status : String = ""
+    private var status: String = ""
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_tv_list)
+        tvListBinding = ActivityTvListBinding.inflate(layoutInflater)
+        setContentView(tvListBinding.root)
         injectDependency()
-        presenter.attach<Activity>(this,this)
+        presenter.attach<Activity>(this, this)
 
         status = intent.getStringExtra(Constant.INTENT_STATUS) ?: ""
 
@@ -50,13 +53,13 @@ class TvListActivity : BaseActivity() ,TvListContract.View{
         setRecycleView()
         setUI()
         if (isConnection)
-        presenter.loadData(currentPage,status)
+            presenter.loadData(currentPage, status)
 
     }
 
     override fun setRecycleView() {
         val mNoOfColumns = Util.calculateNoOfColumns(this)
-        layoutManager = GridLayoutManager(this,mNoOfColumns)
+        layoutManager = GridLayoutManager(this, mNoOfColumns)
         layoutManager.spanSizeLookup = object : GridLayoutManager.SpanSizeLookup() {
             override fun getSpanSize(position: Int): Int {
                 return when {
@@ -72,11 +75,11 @@ class TvListActivity : BaseActivity() ,TvListContract.View{
                 }
             }
         }
-        rvTV.layoutManager = layoutManager
+        tvListBinding.rvTV.layoutManager = layoutManager
     }
 
     override fun setScrollRecycleView() {
-        rvTV.addOnScrollListener(object : PaginationScrollListener(layoutManager){
+        tvListBinding.rvTV.addOnScrollListener(object : PaginationScrollListener(layoutManager) {
             override fun loadMoreItems() {
                 isLoading = true
                 currentPage += 1
@@ -97,9 +100,9 @@ class TvListActivity : BaseActivity() ,TvListContract.View{
 
             override fun backToTop(isShow: Boolean) {
                 if (isShow)
-                    btnBackToTop.show()
+                    tvListBinding.btnBackToTop.show()
                 else
-                    btnBackToTop.hide()
+                    tvListBinding.btnBackToTop.hide()
             }
 
         })
@@ -107,13 +110,13 @@ class TvListActivity : BaseActivity() ,TvListContract.View{
 
     override fun showDataSuccess(list: MutableList<MovieItem?>) {
         countData = list.size
-        if (currentPage == 1){
+        if (currentPage == 1) {
             adapter = TvListAdapter(
-                this,list
+                this, list
             )
-            rvTV.adapter = adapter
+            tvListBinding.rvTV.adapter = adapter
             setScrollRecycleView()
-        }else{
+        } else {
             adapter.removeLoadingFooter()
             isLoading = false
             adapter.addData(list)
@@ -133,45 +136,45 @@ class TvListActivity : BaseActivity() ,TvListContract.View{
         checkData()
     }
 
-    override fun moveToDetail(id : Int){
+    override fun moveToDetail(id: Int) {
         val intent = Intent(this, TvDetailActivity::class.java)
-        intent.putExtra(Constant.idMovie,id)
+        intent.putExtra(Constant.idMovie, id)
         startActivity(intent)
     }
 
     override fun showDataFailed(message: String) {
-        Log.d("ErrorTvList",message)
+        Log.d("ErrorTvList", message)
         showData(false)
     }
 
     override fun checkLastData() {
-        if (countData == limit){
+        if (countData == limit) {
             adapter.addLoadingFooter()
-        }else{
+        } else {
             isLastPage = true
         }
     }
 
     override fun checkData() {
-        if (adapter.itemCount == 0){
+        if (adapter.itemCount == 0) {
             showData(false)
-        }else{
+        } else {
             showData(true)
         }
     }
 
     override fun showData(isShow: Boolean) {
-        if (isShow){
-            rvTV.visibility  = View.VISIBLE
-            noData.visibility   = View.GONE
-        }else{
-            rvTV.visibility  = View.GONE
-            noData.visibility   = View.VISIBLE
+        if (isShow) {
+            tvListBinding.rvTV.visibility = View.VISIBLE
+            tvListBinding.noData.visibility = View.GONE
+        } else {
+            tvListBinding.rvTV.visibility = View.GONE
+            tvListBinding.noData.visibility = View.VISIBLE
         }
     }
 
-    override fun getTitle(title: String): String? {
-        return when (title){
+    override fun getTitle(title: String): String {
+        return when (title) {
             getString(R.string.tv_airing_today) -> {
                 getString(R.string.tv_airing_today)
             }
@@ -196,14 +199,17 @@ class TvListActivity : BaseActivity() ,TvListContract.View{
     }
 
     override fun setToolbar() {
-        setSupportActionBar(toolbar)
+        setSupportActionBar(tvListBinding.toolbar)
         supportActionBar?.setDisplayHomeAsUpEnabled(true)
-        Util.setColorFilter(toolbar.navigationIcon!!, ContextCompat.getColor(this, R.color.iconColorPrimary))
+        Util.setColorFilter(
+            tvListBinding.toolbar.navigationIcon!!,
+            ContextCompat.getColor(this, R.color.iconColorPrimary)
+        )
         title = getTitle(status)
     }
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        if (item.itemId == android.R.id.home){
+        if (item.itemId == android.R.id.home) {
             onBackPressed()
         }
         return super.onOptionsItemSelected(item)
@@ -220,26 +226,28 @@ class TvListActivity : BaseActivity() ,TvListContract.View{
     }
 
     override fun setUI() {
-        btnBackToTop.hide()
-        btnBackToTop.setOnClickListener { rvTV.smoothScrollToPosition(0) }
-        refreshLayout.setOnRefreshListener {
-            refreshLayout.isRefreshing = false
+        tvListBinding.btnBackToTop.hide()
+        tvListBinding.btnBackToTop.setOnClickListener { tvListBinding.rvTV.smoothScrollToPosition(0) }
+        tvListBinding.refreshLayout.setOnRefreshListener {
+            tvListBinding.refreshLayout.isRefreshing = false
             currentPage = 1
-            presenter.loadData(currentPage,status)
+            presenter.loadData(currentPage, status)
         }
     }
 
     override fun showLoading() {
-        shimmerView.visibility  = View.VISIBLE
-        shimmerView.setShimmer(Shimmer.AlphaHighlightBuilder().setDuration(1150L).build())
-        shimmerView.startShimmer()
-        refreshLayout.visibility = View.GONE
+        tvListBinding.shimmerView.visibility = View.VISIBLE
+        tvListBinding.shimmerView.setShimmer(
+            Shimmer.AlphaHighlightBuilder().setDuration(1150L).build()
+        )
+        tvListBinding.shimmerView.startShimmer()
+        tvListBinding.refreshLayout.visibility = View.GONE
     }
 
     override fun hideLoading() {
-        shimmerView.stopShimmer()
-        shimmerView.visibility      = View.GONE
-        refreshLayout.visibility    = View.VISIBLE
+        tvListBinding.shimmerView.stopShimmer()
+        tvListBinding.shimmerView.visibility = View.GONE
+        tvListBinding.refreshLayout.visibility = View.VISIBLE
     }
 
 }
