@@ -7,6 +7,7 @@ import android.view.ViewGroup
 import android.view.WindowManager
 import androidx.core.content.ContextCompat
 import id.fajarproject.roommovie.R
+import id.fajarproject.roommovie.databinding.ActivityVideoBinding
 import id.fajarproject.roommovie.di.component.DaggerActivityComponent
 import id.fajarproject.roommovie.di.module.ActivityModule
 import id.fajarproject.roommovie.ui.base.BaseActivity
@@ -14,20 +15,21 @@ import id.fajarproject.roommovie.util.Constant
 import id.fajarproject.roommovie.util.LoadingWebViewClient
 import id.fajarproject.roommovie.util.Util
 import id.fajarproject.roommovie.util.VideoWebChromeClient
-import kotlinx.android.synthetic.main.activity_video.*
 
 class VideoPlayerActivity : BaseActivity() {
 
-    private var keyVideo : String? = null
+    private var keyVideo: String? = null
+    private lateinit var videoBinding: ActivityVideoBinding
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_video)
+        videoBinding = ActivityVideoBinding.inflate(layoutInflater)
+        setContentView(videoBinding.root)
         injectDependency()
         setToolbar()
-        title    = intent.getStringExtra(Constant.title)
+        title = intent.getStringExtra(Constant.title)
         keyVideo = intent.getStringExtra(Constant.keyVideo)
-        if (keyVideo?.isNotEmpty() == true){
+        if (keyVideo?.isNotEmpty() == true) {
             setViewVideo()
         }
     }
@@ -40,32 +42,38 @@ class VideoPlayerActivity : BaseActivity() {
         activityComponent.inject(this)
     }
 
-    fun setToolbar(){
-        setSupportActionBar(toolbar)
+    fun setToolbar() {
+        setSupportActionBar(videoBinding.toolbar)
         supportActionBar?.setDisplayHomeAsUpEnabled(true)
-        Util.setColorFilter(toolbar.navigationIcon!!, ContextCompat.getColor(this, R.color.iconColorPrimary))
+        Util.setColorFilter(
+            videoBinding.toolbar.navigationIcon!!,
+            ContextCompat.getColor(this, R.color.iconColorPrimary)
+        )
 
     }
 
-    private fun setViewVideo(){
+    private fun setViewVideo() {
         val nonVideoLayout: View =
             findViewById(R.id.videoLayout)
 
         val videoLayout: ViewGroup =
             findViewById(R.id.videoFullScreenOverLay)
 
-        ivThumbnail.visibility  = View.GONE
-        videoPlayer.visibility  = View.VISIBLE
-        progressBar.visibility  = View.GONE
+        videoBinding.ivThumbnail.visibility = View.GONE
+        videoBinding.videoPlayer.visibility = View.VISIBLE
+        videoBinding.progressBar.visibility = View.GONE
 
         //noinspection all
         val loadingView = layoutInflater.inflate(
             R.layout.custom_loading_video,
-            null)
+            null
+        )
 
-        val webChromeClient                 = VideoWebChromeClient(nonVideoLayout,videoLayout,loadingView,videoPlayer)
+        val webChromeClient =
+            VideoWebChromeClient(nonVideoLayout, videoLayout, loadingView, videoBinding.videoPlayer)
 
-        webChromeClient.setOnToggledFullscreen(object : VideoWebChromeClient.ToggledFullscreenCallback{
+        webChromeClient.setOnToggledFullscreen(object :
+            VideoWebChromeClient.ToggledFullscreenCallback {
             override fun toggledFullscreen(fullscreen: Boolean) {
                 // Your code to handle the full-screen change, for example showing and hiding the title bar. Example:
                 if (fullscreen) {
@@ -75,7 +83,7 @@ class VideoPlayerActivity : BaseActivity() {
                     attrs.flags = attrs.flags or WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON
                     window.attributes = attrs
                 } else {
-                    requestedOrientation = ActivityInfo.SCREEN_ORIENTATION_PORTRAIT
+                    requestedOrientation = ActivityInfo.SCREEN_ORIENTATION_UNSPECIFIED
                     val attrs = window.attributes
                     attrs.flags = attrs.flags and WindowManager.LayoutParams.FLAG_FULLSCREEN.inv()
                     attrs.flags =
@@ -84,11 +92,17 @@ class VideoPlayerActivity : BaseActivity() {
                 }
             }
         })
-        videoPlayer.webChromeClient = webChromeClient
-        videoPlayer.webViewClient   = LoadingWebViewClient(this)
-        videoPlayer.settings.mediaPlaybackRequiresUserGesture = false
+        videoBinding.videoPlayer.webChromeClient = webChromeClient
+        videoBinding.videoPlayer.webViewClient = LoadingWebViewClient(this)
+        videoBinding.videoPlayer.settings.mediaPlaybackRequiresUserGesture = false
 
-        videoPlayer.loadDataWithBaseURL("http://www.youtube.com", Util.getScriptAutoPlayYoutube(keyVideo), "text/html", "utf-8", null)
+        videoBinding.videoPlayer.loadDataWithBaseURL(
+            "http://www.youtube.com",
+            Util.getScriptAutoPlayYoutube(keyVideo),
+            "text/html",
+            "utf-8",
+            null
+        )
 //        videoPlayer.loadUrl(Constant.BASE_VIDEO + keyVideo + "?autoplay=1")
     }
 
